refactor api service for easy extensibility
