+++ src/main/scala/krs/api/APIServer.scala
@@ -8,51 +8,27 @@ import com.twitter.server.TwitterServer
 
 import io.finch._
 import io.finch.circe._
-import io.circe.{ Encoder, Json }
-
-import com.twitter.finagle.Thrift
 
 import com.twitter.util.{ Await }
 
-import krs.thriftscala.{ PartnerService, OfferResponse }
-
-trait PartnerResponseEncoders {
-  implicit val encodeOfferResponse: Encoder[OfferResponse] = Encoder.instance(e =>
-    Json.obj(
-      "offers" -> Json.fromValues(e.offers.map((o) => {
-        Json.obj(
-          "provider" -> Json.fromString(o.provider),
-          "minimumScore" -> Json.fromInt(o.minimumCreditScore.getOrElse(0)),
-          "maximumScore" -> Json.fromInt(o.maximumCreditScore.getOrElse(0))
-        )
-      }))
-    )
-  )
+object APIService extends OfferResponseEncoders {
+  val api: Service[Request, Response] = (
+    OfferAPI.getOffers
+  ).toServiceAs[Application.Json]
 }
 
-object APIServer extends TwitterServer with PartnerResponseEncoders {
+object APIServer extends TwitterServer {
 
   val port: Flag[Int] = flag("port", 8080, "TCP port for HTTP server")
 
-  val todos: Counter = statsReceiver.counter("partnerclient")
-
-  val client: PartnerService.FutureIface =
-    Thrift.client.newIface[PartnerService.FutureIface]("localhost:8081", classOf[PartnerService.FutureIface])
-
-  def getOffers: Endpoint[OfferResponse] = get("offers") {
-    client.getOffers().map(Ok)
-  }
-
-  val api: Service[Request, Response] = (
-    getOffers
-  ).toServiceAs[Application.Json]
+  val api: Counter = statsReceiver.counter("api")
 
   def main(): Unit = {
     log.info("Serving the Client application")
 
     val server = Http.server
       .withStatsReceiver(statsReceiver)
-      .serve(s":${port()}", api)
+      .serve(s":${port()}", APIService.api)
 
     onExit { server.close() }
